node/src/node/requirements.txt
# Matches the recent tag line but tolerates PyPI lag
crewai>=0.165,<0.178
crewai-tools>=0.62,<0.70

# your other deps
pydantic>=2.5,<3
python-dotenv>=1.0,<2
# Language model providers
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.10

# Essential tools and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File and directory operations
pathlib2>=2.3.7
watchdog>=3.0.0

# Development tools
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.5.0

# Testing and quality assurance
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP and API interactions
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.0

# JSON and YAML processing
pyyaml>=6.0.1
jsonschema>=4.19.0

# Process and subprocess management
psutil>=5.9.0

# Logging and monitoring
loguru>=0.7.0
colorlog>=6.7.0

# Security scanning
bandit>=1.7.5
safety>=2.3.5

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0

# Optional: If using OpenAI
openai>=1.0.0

# Optional: If using Anthropic directly
anthropic>=0.8.0

# Optional: For web search capabilities
serpapi>=0.1.5
googlesearch-python>=1.2.3

# Optional: For code analysis
astroid>=3.0.0
pylint>=3.0.0
radon>=6.0.0

# Version management
packaging>=23.1
semver>=3.0.1